+++ src/test/java/kofa/colours/spaces/Rec2020Test.java
@@ -11,51 +11,51 @@ class Rec2020Test {
     @Test
     void rec2020_to_XYZ_and_back() {
         // given
-        float[] rec2020 = new float[] {1, 1, 1};
-        float[] XYZ = newBuffer();
+        double[] rec2020 = new double[] {1, 1, 1};
+        double[] XYZ = newBuffer();
 
         // when
         Rec2020.rec2020_to_XYZ(rec2020, XYZ);
 
         // then
-        assertThat(XYZ[0]).isCloseTo((float) D65_WHITE_2DEGREE_STANDARD_OBSERVER.X(), offset(1e-6f));
-        assertThat(XYZ[1]).isCloseTo((float) D65_WHITE_2DEGREE_STANDARD_OBSERVER.Y(), offset(1e-6f));
-        assertThat(XYZ[2]).isCloseTo((float) D65_WHITE_2DEGREE_STANDARD_OBSERVER.Z(), offset(1e-6f));
+        assertThat(XYZ[0]).isCloseTo(D65_WHITE_2DEGREE_STANDARD_OBSERVER.X(), offset(1e-6));
+        assertThat(XYZ[1]).isCloseTo(D65_WHITE_2DEGREE_STANDARD_OBSERVER.Y(), offset(1e-6));
+        assertThat(XYZ[2]).isCloseTo(D65_WHITE_2DEGREE_STANDARD_OBSERVER.Z(), offset(1e-6));
 
         // when
         Rec2020.XYZ_to_rec2020(XYZ, rec2020);
 
         // then
-        assertThat(rec2020[0]).isCloseTo(1, offset(1e-6f));
-        assertThat(rec2020[1]).isCloseTo(1, offset(1e-6f));
-        assertThat(rec2020[2]).isCloseTo(1, offset(1e-6f));
+        assertThat(rec2020[0]).isCloseTo(1, offset(1e-6));
+        assertThat(rec2020[1]).isCloseTo(1, offset(1e-6));
+        assertThat(rec2020[2]).isCloseTo(1, offset(1e-6));
     }
 
     @Test
     void rec2020_and_rec709() {
         // given values picked in darktable
-        float[] rec2020 = new float[] {101 / 255.0f, 114 / 255.0f, 170 / 255.0f};
-        float[] valuesXYZ = newBuffer();
+        double[] rec2020 = new double[] {101 / 255.0, 114 / 255.0, 170 / 255.0};
+        double[] valuesXYZ = newBuffer();
 
         // when
         Rec2020.rec2020_to_XYZ(rec2020, valuesXYZ);
         // then
-        assertThat(valuesXYZ[0]).isCloseTo(0.42950739766460216f, offset(1e-6f));
-        assertThat(valuesXYZ[1]).isCloseTo(0.4466875440341956f, offset(1e-6f));
-        assertThat(valuesXYZ[2]).isCloseTo(0.719768499718504f, offset(1e-6f));
+        assertThat(valuesXYZ[0]).isCloseTo(0.42950739766460216, offset(1e-6));
+        assertThat(valuesXYZ[1]).isCloseTo(0.4466875440341956, offset(1e-6));
+        assertThat(valuesXYZ[2]).isCloseTo(0.719768499718504, offset(1e-6));
 
         // when
-        float[] rec709 = newBuffer();
+        double[] rec709 = newBuffer();
         Rec709.XYZ_to_rec709(valuesXYZ, rec709);
 
         // then we get values picked in darktable
         // rough comparison, 8-bit values
-        assertThat(rec709[0]).isCloseTo(89 / 255f, offset(1e-2f));
-        assertThat(rec709[1]).isCloseTo(115 / 255f, offset(1e-3f));
-        assertThat(rec709[2]).isCloseTo(177 / 255f, offset(1e-3f));
+        assertThat(rec709[0]).isCloseTo(89 / 255, offset(1e-2));
+        assertThat(rec709[1]).isCloseTo(115 / 255, offset(1e-3));
+        assertThat(rec709[2]).isCloseTo(177 / 255, offset(1e-3));
         // exact values based on other code
-        assertThat((rec709[0])).isCloseTo(0.3464098892162477f, offset(1e-6f));
-        assertThat((rec709[1])).isCloseTo(0.45157499402975027f, offset(1e-6f));
-        assertThat((rec709[2])).isCloseTo(0.6936708110984687f, offset(1e-6f));
+        assertThat((rec709[0])).isCloseTo(0.3464098892162477, offset(1e-6));
+        assertThat((rec709[1])).isCloseTo(0.45157499402975027, offset(1e-6));
+        assertThat((rec709[2])).isCloseTo(0.6936708110984687, offset(1e-6));
     }
 }
